Update Simple Calculator - terminal
Is a simple computer application made in java. The user can see in the terminal the main menu containing the available operators (+,-,/,*,^). The user is asked in turn the first element the operator and the second element and displays the result. To make the interaction with the user more pleasant, we used ANSI codes for the color of the text.